+++ pages/instructorAssessment/instructorAssessment.js
@@ -286,6 +286,13 @@ router.get('/:filename', function(req, res, next) {
                 ['Correct', 'correct'],
                 ['Feedback', 'feedback'],
             ];
+            if (!include_all) {
+                columns = columns.concat([
+                    ['Question points', 'points'],
+                    ['Max points', 'max_points'],
+                    ['Question % score', 'score_perc'],
+                ]);
+            }
             csvMaker.rowsToCsv(result.rows, columns, function(err, csv) {
                 if (ERR(err, next)) return;
                 res.attachment(req.params.filename);